Move test cases to right place
Because migrationAccounts is method of Wallet class,
it must be placed at WalletAPITest, not KASWalletTest.
